added offset pattern of isLoadingScreen

+++ src/Program.cs
@@ -117,7 +117,7 @@ namespace PoeHUD
                 sb.Append(System.Environment.NewLine);
 
 
-                sb.Append("Area change: " + memory.ReadLong(offs.AreaChangeCount + memory.AddressOfProcess));
+                sb.Append("Area change: " + memory.ReadInt(offs.AreaChangeCount + memory.AddressOfProcess));
                 sb.Append(System.Environment.NewLine);
                 sb.Append(System.Environment.NewLine);
